pyproject.toml
[tool.poetry]
name = "cloudlet-ai-backend"
version = "0.1.0"
description = ""
authors = ["Oliver Hulland <olihulland@gmail.com>"]
readme = "README.md"
packages = [{include = "cloudlet_ai_backend"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
flask = "^2.3.2"
pyserial = "^3.5"
flask-cors = "^3.0.10"
flask-socketio = "^5.3.4"
simple-websocket = "^0.10.0"
tensorflow-macos = "^2.12.0"
tensorflow-hub = "^0.13.0"

[tool.poetry.scripts]
start = "cloudlet_ai_backend.main:start_dev"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
